Building out the typename facility to register member functions associated with a type. Constructor functions now automatically register objects returned for destruction, associated with the VM's current call frame. Doubles changed to floats.

+++ Compiler/FloatConstantExpression.h
@@ -5,7 +5,7 @@
 class FloatConstantExpression : public Expression
 {
 protected:
-	double mValue;
+	float mValue;
 
 public:
 	FloatConstantExpression(int rootToken)
